feat(cart): normalize CartItem.shippingCost to double? with safe JSON parsing
- Change CartItem.shippingCost from int? to double? to accommodate both int and double payloads.
- Parse as (num?)?.toDouble() in fromJson.
- Verified via project-wide search that CartItem.shippingCost is not referenced elsewhere (model-only field).

+++ lib/data_model/cart_response.dart
@@ -81,7 +81,7 @@ class CartItem {
   String? price;
   String? currencySymbol;
   String? tax;
-  int? shippingCost;
+  double? shippingCost;
   int quantity;
   int? lowerLimit;
   int upperLimit;
